app/requirements.txt
sqlalchemy==2.0.31
sqlmodel

asyncpg==0.28.0
psycopg2-binary
psycopg-binary==3.1.10

alembic==1.12.0
pydantic==2.3.0
pydantic-settings==2.0.3

fastapi==0.109.0
uvicorn
starlette
bcrypt

python-jose==3.4.0

python-multipart
pika
transformers

jinja2
